[skip ci] adding Teal's changes to primenet
