+++ .gitignore
@@ -1,1 +1,2 @@
 .ipynb_checkpoints/
+*.idea
